created common lib
