+++ src/common/constants.py
@@ -1,0 +1,2 @@
+MAC_OS = "darwin"
+
